Refactor Clone Tests helpers
